исправлены замечания

+++ src/main/java/ru/skypro/homework/dto/CommentDto.java
@@ -16,7 +16,7 @@ public record CommentDto(
         /**
          дата и время создания комментария в миллисекундах с 00:00:00 01.01.1970
          */
-        Integer createdAt,
+        Long createdAt,
         /**
          id комментария
          */
